Flask API/requirements.txt
Flask==3.1.1
python-dotenv==1.0.1
Werkzeug==3.1.3
supabase==2.5.1
gotrue==2.5.0
paho-mqtt==2.1.0
sendgrid==6.11.0
firebase-admin==6.5.0
schedule==1.2.2